Add tm yakbak tests
