+++ api/src/services/tm_types/index.js
@@ -1,7 +1,9 @@
 const TM2API = require('./tm2')
 const TM3API = require('./tm3')
+const FakeTMAPI = require('./fake_tm')
 
 module.exports = {
   TM2API,
-  TM3API
+  TM3API,
+  FakeTMAPI
 }
